moves functions to shared files
