+++ packages/reports/src/utils/js/sorter.js
@@ -1,0 +1,1 @@
+module.exports = (a, b) => a.ordering - b.ordering
